:truck: move 'assert_python' out of main.py
